First set of changes to make DRLVM work with the launcher from the perspective of loading libraries and resolving library dependencies.
Main problem was that all libraries were specified with full
path, and APR just passes through to dlopen(), which will not
respect LD_LIBRARY_PATH if a fully qualified path is used.

So these changes remove qualification - and we can now
depend on the Harmony launcher to establish LD_LIBRARY_PATH
with the proper directory.

It was necessary to also set the DRLVM-specific property
vm.boot.library.path to "" rather than directory of executable
not only for this purpose, but because given the DRLVM can be
instantiated in many ways, this seemed impractical.

Finally, build is slightly modified to create a libharmonyvm.so
rather than libvmcore.so as per the recent changes to the 
Harmony launcher.


svn path=/incubator/harmony/enhanced/drlvm/trunk/; revision=440956