Next step in the move to full launcher compatibility.
There are still some internal problems, but structurally
we work.  The build dumps into a /default directory, the 
help and -X stuff are fixed, and it seems to not fall over
when running code.  

This has been tested on Ubuntu 6 - once checked in, I'll test
and fix as necessary on Win32.

Lots of this either inspired by or taken from HARMONY-1390, so
thanks to Evgueni Brevnov for blazing the trail.


svn path=/incubator/harmony/enhanced/drlvm/trunk/; revision=441745