HARMONY-1427
Looks like the mode in which exn_print_stack_trace works has changed. It is now suspend_enabled, and inside of this 
function there is an assertion that suspend is enabled. But the change was not applied to JNI ExceptionDescribe 
function. It still disables suspend when calling exn_print_stack_trace.

tested on ubuntu 6, smoke and c-unit ok, 


svn path=/incubator/harmony/enhanced/drlvm/trunk/; revision=446796