HARMONY-1430
The bug in optimizing simple arithmetic instruction consist of two parts:
    - copy/paste caused bug in simplifier::simplifyAnd(xxx)
    - erroneuos conversion elimination in abcd
The first is fixed in the code.
The second is fixed by turning off conversion elimination in abcd.
(It should be rewritten completely later)


The second bug: Jitrino does not throw a LinkingException when ldc tries to load
incorrect class literal.

As 1.5 support is available 'ldc' can loads class literals.
These classes may be incorrect (not resolved successfully).
It actually does not distrub ldc activity but according to specification
LinkingException must be thrown.
Bugfix:ThrowLinkingException helper call is generated at the place of ldc in managed
code if literal type can not be resolved. 



svn path=/incubator/harmony/enhanced/drlvm/trunk/; revision=446800