HARMONY-1442
While TI is on, VM crashes when it attempts to define class which contains
method with absent local variable table information through JNI DefineClass.

When JVMTI is off, the local variable table of a class is not parsed, just skipped. When JVMTI is on, the local 
variable table was parsed just fine. But the ambigous interface treated NULL returned by the parsing function as an 
error, regardless of whether the local variable table was incorrect or just absent. Usually local variable table is 
present, but classes are allowed not to have it. So when this table was just absent in a class, an error was returned 
just like parsing was failed.


tested on ubuntu - smoke and c-unit passed



svn path=/incubator/harmony/enhanced/drlvm/trunk/; revision=447080