HARMONY-1455
The implementation of SingleStep even is based on Breakpoints support in JVMTI and HARMONY-1422 next bytecode 
predictor.

This actual patch contains initial SingleStep support added to the breakpoints handler of JVMTI. It also 
enableds/disables SingleStepping when SingleStepping is turng on/off. It should start SingleStepping exception 
handler and when running a new Java method from VM native code. When execution control is transferred to Java, it 
should start SingleStepping it from the transfer location.

TODO is to check that this implementation actually works 

Tested on ubuntu  using smoke and c-unit


svn path=/incubator/harmony/enhanced/drlvm/trunk/; revision=447224