HARMONY-1517
jvmtiGetThreadGroupChildren method returns threads which belongs not only
to the given group but to all children groups recurcively, which contradicts
to the spec.

Also jvmtiGetAllThreads method according to the spec should return only alive
ones, while currently it returns threads which was not yet run.

committed and tested on Ubuntu 6 - note that I have confidence only in c-unit and smoke.  There are
failing tests in kernel, but they seem to always have been


svn path=/incubator/harmony/enhanced/drlvm/trunk/; revision=448645