HARMONY-1575
Single stepping back from java to native requres predicting a next bytecode after *return one. But there is an 
assertion that frame is java. If the frame is native it fails. This assertion was removed because it cannot be 
satisfied always. But instead I removed incorrect assumption comments on returning from native code back to Java (it 
it returns to Java at all) to start single stepping back in Java code. If the calling code was not Java single 
stepping won't happen, it will just wait another native from Java transition.

Ubuntu 6 - smoke, c-unit and kernel


svn path=/incubator/harmony/enhanced/drlvm/trunk/; revision=450709