HARMONY-1123
computeSerialVersionUID() returned the wrong value compared to
the scribbled footnote jokingly referred to as the spec as
implemented by the RI

The result was a serialized instance of a class that didn't have
a set serialVersionUID field wasn't de-serializable by Harmony.

For simple tests, we now return the same ID as the RI.

All classlib tests passed.

We need more tests, ones that specifically test this with complicated
classes that are precompiled with the RI's serialVersionUID known.



svn path=/incubator/harmony/enhanced/classlib/trunk/; revision=451082