HARMONY-1613
In interpreter mode function GetStackTrace() misses frame with main() function.
The problem is in skip_old_frames() function in interpreter. It reports wrong additional missing frame.

Wasn't able to test this specific fix, but it didn't break anything.  Can we have a test to show it?

Ubuntu 6, smoke, c-unit and kernel (as they are)


svn path=/incubator/harmony/enhanced/drlvm/trunk/; revision=452152