HARMONY-1651
Mostly - rejected the catch(NPE) and did something to be explicit
about what we are doing when there is no specVersion.

All three tests pass.  Note, I had to also add a manifest to the
kernel.jar for this to work - not sure how this worked without it.

All three tests pass, and nothing else seems to have broken.

Ubuntu 6, smoke, c-unit, ~kernel


svn path=/incubator/harmony/enhanced/drlvm/trunk/; revision=453042