HARMONY-1561
Interpreter throws IllegalAccessError on putfield opcode for all methods except <init>.
It should not throw the error if the <init> method for this object is in the calling stack.
Reproducible on activemq-4.0.2.

Had to -R putfield.patch before applying H-1561.diff

TestPutFinal.j passes , as do smoke, c-unit, kernel on Ubuntu 6



svn path=/incubator/harmony/enhanced/drlvm/trunk/; revision=463912