Applied HARMONY-2721 [drlvm][test] Add test in HARMONY-2083 to regression test stuite
It appears that there is a race condition in using class_report_failure because
it is called after Class is unlocked. So it is possible that 2 threads come
with different errors to calling cp.resolve_as_error. To avoid errors
duplication since it is necessary to use only the first error it is better
to check for existing error before calling cp.resolve_as_error.


svn path=/harmony/; revision=499839