Disallow in class_find_by_cp_entry() all constant_pool entries except UTF8.
Added support for reading class info from Java
archives as well as explicit class files.

Added test for an abstract method in class_load_resolve_clinit()
with same arbitrary decision as for when non 'clinit' method
was available, and marked as a to-do item.

Mark JVM for shutdown if no previous thread to revive.

svn path=/harmony/enhanced/sandbox/; revision=565142