Fixed HARMONY-5238 [drlvm][util] memory leak at compile_gen_compile_me at compile_ia32.cpp while EIOffice run
Leaked name is freed now, those names that point to static data are not freed. It was
necessary to add a flag to DynamicCode struct to distinguish between these two kinds of
storage for the name.

Also I enabled dynamic code collection for compile_me stubs in compile_em64t


svn path=/harmony/; revision=601355