Combine the strLen calculation and finding loops to optimise for the highly likely case that both will happen.
Fix off-by-one error that occurs because the strLen calculation and
the separator concatenation logic were not consistent.  This occurred
if you did something like:

  LD_LIBRARY_PATH=$PWD/deploy/jdk/jre/bin/default deploy/jdk/jre/bin/java ...

where the first newPathToAdd is missing but the second is found
leaving strLen with the wrong number of separators.  (It occurred for
me because the first newPathToAdd requires a trailing slash while the
second does not and I provided both without slashes.  We should fix
this too.)

svn path=/harmony/; revision=820201