Fixing the bugs demonstrated by WriterTesterTest. Only ObjectOutputStream failures are outstanding; these are suppressed in the test case.
This commit adds SneakyThrow, which is a dangerous class that you should not use! It is used to give Harmony the best behaviour when dealing with failures during stream flush and close. Read the Javadoc for full detalis.

This commit makes behaviour changes to BufferedWriter, FilterOutputStream, and BufferedOutputStream. These streams now throw as required after they are closed, and throw the correct exception when closing fails.

svn path=/harmony/; revision=835972