Remove jars that haven't been used by trunk for over a year. Remove stubs and build.xml that are no longer needed.
Add new jar to enable testing (not using a directory since it isn't
really necessary since the upstream jar name contains the version).

Testing diff for make/depends.properties would be something like:

-bcprov.ver=bcprov-jdk15-144
+bcprov.ver=bcprov-jdk15-145
 bcprov.dir=${depends.jars}/${bcprov.ver}
 bcprov.jar=${bcprov.dir}/bcprov.jar
-bcprov.url=${depends.url}/jars/bcprov/${bcprov.ver}/bcprov.jar
-bcprov.md5=0d6f0b31730db6fd1f8360cf1f7666d4
+bcprov.url=${depends.url}/jars/bcprov/${bcprov.ver}.jar
+bcprov.md5=2062f8e3d15748443ea60a94b266371c

Note that the copyright also needs to be updated in several places if/when
this is committed.


svn path=/harmony/; revision=900350